Autoterminal: Split terminal logic & font drawing
